Completed Chat Refactor from New Dev Branch
Replaced Polls with Chats, moved all models except message to gostem app, chat app is functional
