+++ backend/chats/routing.py
@@ -1,5 +1,5 @@
 from django.urls import re_path 
-from polls.consumers import ChatConsumer
+from chats.consumers import ChatConsumer
 
 websocket_urlpatterns = [
     re_path(r'ws/chat/$', ChatConsumer.as_asgi()),
