+++ backend/chats/views.py
@@ -1,8 +1,8 @@
 from rest_framework import status
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
-from .models import User, Event
-from .serializers import UserSerializer, EventSerializer
+from gostem.models import User, Event
+from gostem.serializers import UserSerializer, EventSerializer
 from django.shortcuts import render
 
 @api_view(['POST'])
@@ -21,5 +21,6 @@ def create_event(request):
         return Response(serializer.data, status=status.HTTP_201_CREATED)
     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-def chatPage(request):
-    return render(request, "polls/chat.html")
+def chatPage(request, *args, **kwargs):
+    context = {}
+    return render(request, "chats/chatPage.html", context)
