+++ backend/gostem/settings.py
@@ -37,13 +37,13 @@ INSTALLED_APPS = [
     "django.contrib.contenttypes",
     "django.contrib.sessions",
     "django.contrib.messages",
-    #'daphne',
+    'daphne',
     "django.contrib.staticfiles",
     'rest_framework',
+    'gostem',
+    'chats',
 
-    'polls',
-
-    #'channels',
+    'channels',
     'notes',
