+++ backend/manage.py
@@ -4,6 +4,8 @@ import os
 import sys
 
 
+sys.path.append(os.path.join(os.path.dirname(__file__), 'backend'))
+
 def main():
     """Run administrative tasks."""
     os.environ.setdefault("DJANGO_SETTINGS_MODULE", "gostem.settings")
